requirements.txt
torch>=2.0.0
numpy>=1.24.0
gymnasium
